extract event adapter
